refactor: split pages.ts into focused modules (crud, tree, export, index)
Split the 1206-line src/commands/pages.ts into 4 files under
src/commands/pages/: crud.ts (CRUD handlers + buildClient), tree.ts
(tree helpers + copy-tree), export.ts (export logic), and index.ts
(command registration). Pure refactor with no behavior changes.
